+++ src/components/Form/CheckBox/CheckBox.stories.tsx
@@ -9,5 +9,6 @@ export const Default: FictoanStory<typeof CheckBox> = createStoryFromTemplate(Te
 Default.args = {
     label: "Accept terms & conditions",
     id: "accept-checkbox",
-    name: "accept-terms"
+    name: "accept-terms",
+    required: true
 };