+++ src/components/Form/FormWrapper/FormWrapper.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 
 import { Element } from "../../Element/Element";
-import { CommonAndHTMLProps } from "../../Element/constants";
+import { CommonAndHTMLProps, SpacingTypes } from "../../Element/constants";
 
 import { FormWrapperStyled } from "./FormWrapper.styled";
 
 // prettier-ignore
 export interface FormWrapperCustomProps {
-    spacing ? : "none" | "nano" | "micro" | "tiny" | "small" | "medium" | "large" | "huge";
+    spacing ? : SpacingTypes;
 }
 
 export type FormWrapperElementType = HTMLFormElement;