+++ src/components/Form/Select/Select.stories.tsx
@@ -23,5 +23,6 @@ Default.args = {
             label: "...",
             value: "..."
         },
-    ]
+    ],
+    required: true
 };